Remove vote models & migrations to ship to Heroku
Migration in production was failing, and I had planned to swap the 2 vote models (post_vote and comment_vote) for one polymorphic model, so this seemed opportune.
Also took the chance to amend the post and comment creation migrations to exclude vote columns instead of removing them later.
Removed the vote fixtures and unit tests.