Improve Fake module and fully test it
PostMaker now ensures all records it creates have created_at datetimes _after_ the user who posted them, and any other dependencies (e.g. comments are created after the post they're on and their parent comment (if applicable)).

I used Time::at in Fake::creation_date instead of something more idiomatic (eg) Integer.seconds.ago, because it felt more natural for the interface to request datetimes that the creation-date needed to be between, (rather than a range of seconds-ago).

To test the randomly generated dates, I decided to seed rand so that it returns values at the extreme ends of its range. Then I check that the values returned are in the range we care about (e.g. after all the supplied records were created, but before 1 hour ago).

Tests that rely on the quick_random_record gem crash when using Arel to submit a raw SQL requests for the max and min values of id column of a table. This seems strange – if it works in the development SQLite DB, it should work in test. I obviously need to do some research. For the time-being i've decided to use 'strategy: 3', which plucks the primary key and samples it in memory.

Tests for each subredidit are dynamically defined. This keeps code DRY and includes subredidit names in failure messages so they're easier to understand.
Updated guardfile to run all Fake tests when either Fake file is altered.
I renamed the file containing the PostMaker part of Fake module, because 'helpers' was not very descriptive.

I configured the Guardfile to run all tests on the tests in test/lib after changes to either file comprising the Fake module, as well as running all tests on the posts resource after changes in the PostMaker file, in case the changes were to Fake::creation_date.