Update seeds & fixtures (make created_at random)
I decided to put the method to generate the creation datetime in the Fake module, so that it's available to the seeds and fixtures, but it's not autoloaded (which would slow the site in production).

I decided to set updated_at times to the same as created_at times. In the future it may make a difference for showing whether a post/comment has been edited.

I added a method to generate passwords for randomly generated users, to make them slightly more secure. If Heroku config vars could be lambda functions, this would be preferred (since this codebase is open). An alternative could be to set the same password for all users and store it as Heroku config var.