Use factories over fixtures for tests
I added the factory_bot_rails gem (only in the test environment), and configured the test_helper and Guardfile to use the factories.

I made the required adjustments to the way the test records are retrieved in the test_helper and the few tests that retrieve them directly. I also moved the seldom-used @other_user in the test_helper into its own getter method, to reduce overhead.

I made the users_index test run faster by only checking the user summaries until it confirms that the current user (which is currently displayed first, by design), and at least one other user, display on the first page. If in the future the current user happens to be off the page, the test will fail, and may need updating; the failure message explains this.