Update tests to use factories
After implementing factories for tests, the test suite was passing with many false positives becaused my local test DB was unexpectedly populated, and these records were being tested. I have Amended all model and integration tests to explicitly create users from factories and test these.

Created directories for the integration tests for each resource and updated the Guardfile accordingly: So that updating files related to the resource still runs all integration tests, and so that when resource's test-helper file ('test/integration/[resource]/[resource]_test.rb') is updated, it runs all the integration tests in that resource's folder.

Moved assertions about non-success flashes into a method in the test helper, and stopped testing for specific codes (eg danger), in case I adopt a different convention in future.

There was a mistake in the :user factory, which was defining the password_digest, rather than the password and password_confirmation. I fixed this.