Display user's posts on their profile page
User feed currently shows all of the user's posts in an unordered list.

Refactored the user-index SCSS class into an extensible to reuse it in the
user index and post feeds.

Used sequences in my FactoryBot factory for posts, so that post titles in tests
are different. This prevents false-negative testing scenarios where displayed
and queried posts are incorrectly matched up.

Also added a sequence for comments, although it is not used yet.