Display metadata above comments
Made some tweaks to the way all metadata is displayed above post/comment
text.

Added a test to check comment text was displaying correctly. It was
initially failing, but eventually I discovered it was because I had
padded the erb with spaces either-side: `<p> <%= foo %> </p>`, and those
spaces were obviously being included. (It's tempting to make this
mistake because erb autocomplete isn't activated if `<` is typed
contiguously with preceding spaces.)