Allow commenting on posts
Currently only possible to post a top-level comment. (You can't reply.)

The post_id is submitted as a hidden form field under the comment
namespace; this conforms to the comment model and makes strong params
simpler. Currently (because comments are ordered recent-first) there's
no need to highlight the just-posted comment, but there will be in
future.

If a user is not logged in, the comment box is replaced by a
login/signup banner, which uses the same partial as the buttons in the
header.