Return to current page after logging in
Clicking a login button on any page returns you to that page once the
login process completes. ApplicationController#Logged_in_user before
filter stores the requested path in the session cookie. When the user is
redirected to sessions/new, the store_location method is called again,
but the cookie isn't overwritten if it's not empty. This could be
avoided by directly rendering the sessions/new template from the
ApplicationController but I felt it wasn't worth violating the
convention of sending a redirect when an unpriveleged user request a
resource.