Seed with more posts and comments
I bumped the number of posts up to 75 so that when subs are online,
there will be enough posts in each to illustrate the ordering in the sub
feed.

I bumped the number of comments in subs where the OP replies to better
illustrate the threaded structure and comment ordering.