Allow users to delete their own comments
Comments with children cannot be deleted per se, instead the text and
the user are redacted, meaning that the children remain. This leaves a
tricky corner case where a parent is deleted (redacted) and then its
only child is deleted. I decided to check for this edge case whenever a
childless comment is deleted, and delete the parent if the case is
found.

Changed variable name post_params => comment_params in
CommentsController. The old name was the result of an unmeticulous copy
+ paste.

Fixed a tiny but fatal mistake in the comment_with_children factory.