Manage comment redaction using callbacks
ActiveRecord wraps all before/after callbacks in transactions to ensure
that if any of them fail, no results will be committed to the DB. For
this reason, redacting a parent comment inside the #cancel_if_parent
callback is impossible. I could have redacted inside an after_rollback
callback instead, but I thought it expressed my intent better to do it
in the controller.