Display threaded comments appropriately
Experimented with creating a comment factory that could create a thread
of comments with dynamic depth, but couldn't get it to work. It seemed
impossible to pass the child parent factory a transient from within the
callback (although i've used this before with static values!). It's as
if the creators specifically wanted to ban recursion.

In the end I just created a grandfather_comment factory and tested the
layout at thread-depths 1, 2 and 3. We may need to revisit this for
very deep threads though.

I removed the default scope on comments because there ARE times I don't
want to use it. I don't think it's sensible to expect coders who DON'T
want something unusual to remember to exclude it; much better to have
those who DO want it ask for it.

This commit also includes an inefficient first attempt at a 'recent'
ordering of comments. I opted for 'recent' to retain thread structures,
ordering each group of siblings (including top level comments as a
group) by age. This means however, that even if a comment is the most
recent on the post, it will still appear in the bottom-most thread if
the comment that started the thread was the first comment on the post.

I decided that a comment should only ever appear once on a page. To
hold with this principle and also allow the most recent comments (with
any ancestry) to rise to the top, one alternative approach occurs:
- Only show recent ancestry
- comments show (eg) only their parent – enough context?!
- a user can click to reveal a comment's hidden ancestry
- if any comments share a displayed ancestry, they are grouped under
  their ancestry and some averaging function sets the effective
  'recentness' of the group

This approach is attractive but too complex for the MVP

The 'recent' ordering has these shortcomings:
- It creates two partial indices in the comments table *_1
- Its performance is bounded by O(n.log(n)) and O(n^2)
- It retrieves parents from the cache even though they're in memory

*_1 One index (parent_id: :asc, created_at: :desc) is incorrect because
the arrangement algorithm in the Comments class works down the list,
moving each comments with a parent_id to the index immediately after
its parent. This means that subsequent (ie earlier) comments with the
same parent_id are moved above the previously moved one. Meaning the
order of all siblings – EXCEPT AT THE TOP LEVEL – is swapped. (The
algorithm MUST work DOWN the list, because if it worked up, it would
move grandchildren to their parents but then move the parents away again
(up the THEIR parents).)

Instead we add a partial index*_2 on the top-level comments:
  `order: { created_at: :desc }, where: 'parent_id = NULL'`
And another on all child comments:
  `order: { parent_id: :asc, created_at: :asc },
   where: 'parent_id NOT NULL'`

*_2 The Comments table was originally indexed by all references. I
deleted these existing indices because none of them will be used in
future, since there will always be some filter applied (eg recent, best,
etc). I modified the existing migration because it makes for a cleaner
history and nobody is downstream.