Introduce voting
Votes are polymorphic; both comments and posts have many of them.
Upvotes are signified by the boolean property :up = true.

Vote routes are shallowly nested under posts and comments.

When a comment or post is deleted its votes are destroyed (even in the
case of a comment with descendents, which will only be anonymised).

Votes are valid if they have a present user and votable (post/comment)
and a boolean up attribute.