Introduce vote controls
Controls available on posts and comments.

Users may only have one vote on any comment or post. But they may change
their vote (eg from up to down). When a post or comment is displayed to
a user who has already voted on it, the voting controls are adjusted to
give them the options of cancelling or reversing their vote.

I created a helper object, VoteContols, to encapsulate the logic
around generating adjusted controls. It makes the views much cleaner.

If a user is using an API to submit a vote, they may have unknowingly
voted already. I decided to support users doing so, by overriding the
build method on the vote association collection, checking for an
existing vote and reversing it as appropriate, or producing an error if
the vote was the same.

To show contextual controls, the the current user's votes on posts and/
or comments must be loaded. It is desirable to avoid loading any other
votes. It becomes challenging to acheive this AND avoid the N + 1 query
problem. For instance, loading all comments on a post, and the user's
votes on all those.

I resolved this conflict by creating a module, VotablePreloading, which
should be mixed into votables' associations with votes, and should
extend votable classes. The module overrides the `includes` method to
force joins on the included associations, joining on votes last, adding
a WHERE clause to restrict to just the user's vote.

There are downsides to this approach:
- Joining can be slower than executing seperate queries, especially for
few results. For a post with many comments, we can expect it to be a
wash.
- The approach cannot be used for multiple models in the same query.
When a user fiews a post, we would like to find the post and preload the
user's vote on it, plus the post's comments, and the user's votes on
THEM. Unfortunately this must be done in two steps: 1 Load post and its
vote; 2 load comments and their votes.

I could alternatively have executed a dynamically created SQL statement
from text, but I would likely have had to sacrifice DB agnosticism to do
so, so it didn't seem worth it.