Prevent users from voting on redacted comments
I refactored the voting controls code because the _vote_controls partial was so
tightly coupled to the VotingControls class in the votable helper that it really
didn't need to exist. I principally prefer avoiding generating html in view
helpers, but in this case it made more sense to handle ALL the business logic in
the helper class and present one interface to the view.