Move vote & comment tests to their own directories
Although comments and post only occur in the context of a post, that's
not a strong reason to consider all their tests sub-tests of a post –
no more than it would to consider tests of all three to be sub-tests of
a user!

Updated guardfile to reflect changes in test location.