Reply to comments
The comments controller now accepts a `comment: :parent_id` parameter,
which is used when replying to manage comment threading.

When a user clicks `reply` beneath a comment, the new-comment form from
the top of the page is copied and emptied/reset by js. I did this in
order to get hold of the required security token, but the method is
quite fragile and opaque. Obviously once the API is up, we can swap to
AJAX.

Use gon gem to pass data from the controller to javascript. Specifically
in this case the ID of the comment that needs to be scrolled to (and
highlighted). Gon is always required in the application layout; views
have to provide gon: true to ensure it is required.

When a user replies to a comment, the post-page is displayed, and their
new comment highlighted and scrolled-to. Because submission is by post
request, and it made more sense to redirect than to render, I needed to
send the comment_ID to the post#show controller somehow. I could have
done this with a session cookie that the controller then deletes, but I
thought it would be better to use a get param, to give users the
(general) ability to save/share direct links to comments. I would be
nice in future to make this more restful... e.g. a request for a comment
routes to the post page that displays it...

If a user bypasses/disables the JS that hides the comment-submit button
and they submit an empty comment, the controller will re-render the
page. But becasue the post request is submitted to comments#create,
refreshing afterwards requests post/1/comments. So I routed that to
posts/1.

Some improvements to the style of error explanations and alerts, but
more needed.

Wherever I touched the style of a component, I also swapped its units of
size to rem (from pt/px) to improve accessibility.